chore: Address comments, add code docs
