--- src/Uno.UWP/Web/Http/CookieSameSiteType.cs
@@ -1,11 +1,0 @@
-﻿#if __WASM__
-namespace Uno.Web.Http
-{
-	public enum CookieSameSite
-	{
-		Lax,
-		Strict,
-		None
-	}
-}
-#endif
